tuning: refactor Optuna into package
