+++ src/lenskit/tuning/_optuna/__init__.py
@@ -1,0 +1,7 @@
+from .results import OptunaTuneResults
+from .search import PipelineTuner
+
+__all__ = [
+    "PipelineTuner",
+    "OptunaTuneResults",
+]
