+++ src/lenskit/tuning/_optuna/search.py
@@ -3,29 +3,23 @@
 # Copyright (C) 2023-2026 Drexel University.
 # Licensed under the MIT license, see LICENSE.md for details.
 # SPDX-License-Identifier: MIT
-
 from __future__ import annotations
 
 import math
-from collections.abc import Iterable
 from concurrent.futures import FIRST_COMPLETED, ThreadPoolExecutor, wait
-from dataclasses import dataclass
 from os import fspath
 
-import numpy as np
 import optuna
 import pandas as pd
 from optuna import Study, Trial
 from optuna.pruners import BasePruner, MedianPruner
 from optuna.storages.journal import JournalFileBackend, JournalStorage
 from optuna.study import StudyDirection
-from optuna.trial import FrozenTrial, TrialState
+from optuna.trial import TrialState
 from pydantic import JsonValue
 from pydantic_core import to_json
 from structlog.stdlib import BoundLogger
-from typing_extensions import override
 
-from lenskit.data import unflatten_dict
 from lenskit.logging import Task, get_logger, item_progress, trace
 from lenskit.logging.tasks import add_context_task
 from lenskit.parallel import NestedPool
@@ -33,12 +27,13 @@ from lenskit.pipeline import Pipeline, PipelineBuilder
 from lenskit.pipeline.components import Component, Placeholder
 from lenskit.pipeline.nodes import ComponentConstructorNode, ComponentInstanceNode
 from lenskit.random import int_seed, make_seed
-from lenskit.schemas.tuning import SearchConfig, SearchParam, SearchSpace, TuningSpec
+from lenskit.schemas.tuning import SearchConfig, SearchParam, SearchSpace
 from lenskit.training import ModelTrainer, TrainingOptions, UsesTrainer
 
-from ._base import BasePipelineTuner, TuneResults
-from ._measure import measure_pipeline
-from ._stopping import PlateauStopRule
+from .._base import BasePipelineTuner
+from .._measure import measure_pipeline
+from .._stopping import PlateauStopRule
+from .results import OptunaTuneResults
 
 _log = get_logger(__name__)
 
@@ -314,7 +309,7 @@ def _ask_space(trial: Trial, space: SearchSpace, *, prefix: str = ""):
             min = int(math.log2(spec.min))
             max = int(math.log2(spec.max))
             p = trial.suggest_int(prefix + name, min, max)
-            out[name] = 2**p
+            out[f"{name}%pow2"] = 2**p
             trace(
                 _log,
                 "sampled int/pow2",
@@ -379,97 +374,12 @@ def _extract_defaults(
                 else:
                     raise e
             if v.scale == "pow2":
-                out[k] = int(math.log2(out[k]))  # type: ignore
+                out[f"{k}%pow2"] = int(math.log2(out[k]))  # type: ignore
+                del out[k]
         else:
             # FIXME this is an ugly hack for single / multiple configs
             out[k] = config
     return out
-
-
-@dataclass
-class OptunaTuneResults(TuneResults):
-    spec: TuningSpec
-    study: Study
-    iterative: bool
-
-    def num_trials(self) -> int:
-        return len([t for t in self.study.trials if t.state != TrialState.FAIL])
-
-    def num_failed(self) -> int:
-        return len([t for t in self.study.trials if t.state == TrialState.FAIL])
-
-    @override
-    def trials(self) -> Iterable[dict[str, JsonValue]]:
-        for trial in self.study.trials:
-            if trial.state != TrialState.FAIL:
-                yield self._trial_result(trial)
-
-    @override
-    def epochs(self) -> Iterable[dict[str, JsonValue]]:
-        if not self.iterative:
-            return []
-
-        for n, trial in enumerate(self.study.trials):
-            if trial.state == TrialState.FAIL:
-                continue
-
-            for i in range(trial.last_step + 1):
-                out_row = {
-                    "trial": n,
-                    "epoch": i,
-                    self.spec.search.metric: trial.intermediate_values.get(i),
-                }
-                yield out_row
-
-    def best_config(self):
-        best = self.study.best_trial
-        cfg = unflatten_dict(best.params)
-        if self.iterative:
-            vals = [best.intermediate_values.get(i) for i in range(best.last_step + 1)]
-            match self.study.direction:
-                case StudyDirection.MAXIMIZE:
-                    cfg["epochs"] = np.argmax(vals).item() + 1
-                case StudyDirection.MINIMIZE:
-                    cfg["epochs"] = np.argmin(vals).item() + 1
-                case _:  # pragma: nocover
-                    raise RuntimeError("unexpected study direction")
-
-        return cfg
-
-    def best_result(self) -> dict[str, JsonValue]:
-        return self._trial_result(self.study.best_trial)
-
-    def _trial_result(
-        self, trial: FrozenTrial, *, include_config: bool = True
-    ) -> dict[str, JsonValue]:
-        assert self.spec.search.metric is not None
-        result: dict[str, JsonValue]
-
-        if self.iterative:
-            assert trial.last_step is not None
-            vals = [trial.intermediate_values.get(i) for i in range(trial.last_step + 1)]
-            epoch = np.argmax(vals).item()
-            result = {
-                self.spec.search.metric: vals[epoch],
-                "epochs": epoch,
-                "attempted_epochs": len(trial.intermediate_values),
-                "final_metric": trial.value,
-            }
-        else:
-            result = {self.spec.search.metric: trial.value}
-
-        if include_config:
-            result["config"] = self._trial_config(trial)
-        return result
-
-    def _trial_config(self, trial: FrozenTrial) -> dict[str, JsonValue]:
-        cfg = unflatten_dict(trial.params)
-        if self.iterative:
-            assert trial.last_step is not None
-            vals = [trial.intermediate_values.get(i) for i in range(trial.last_step + 1)]
-            cfg["epochs"] = np.argmax(vals).item()
-
-        return cfg
 
 
 class CompositePruner(BasePruner):
